+++ noodles-bam/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+
+  * bam/record/codec/encoder/sequence: Encode base as case-insensitive.
+
 ## 0.69.0 - 2024-10-22
 
 ### Changed